requirements.txt
slixmpp
google-generativeai
